+++ catalyst/metrics/__init__.py
@@ -10,8 +10,8 @@ from catalyst.metrics._metric import (
     ICallbackBatchMetric,
     ICallbackLoaderMetric,
     IMetric,
-    AccumulationMetric,
 )
+from catalyst.metrics._accumulative import AccumulativeMetric
 from catalyst.metrics._additive import AdditiveValueMetric
 from catalyst.metrics._confusion_matrix import ConfusionMatrixMetric
 from catalyst.metrics._functional_metric import FunctionalBatchMetric, FunctionalLoaderMetric
